# docs/swagger.yaml
definitions:
  controllers.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controllers.MenuParams:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: string
    type: object
  controllers.MenuRestaurantInput:
    properties:
      menus:
        items:
          $ref: '#/definitions/controllers.MenuParams'
        type: array
      restaurant_id:
        type: string
    type: object
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.RestaurantInput:
    properties:
      city:
        type: string
      name:
        type: string
    type: object
  models.OrderHistory:
    properties:
      created_at:
        type: string
      id:
        type: integer
      restaurantID:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      userID:
        type: integer
    type: object
  models.Restaurant:
    properties:
      city:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /create/menus:
    post:
      consumes:
      - application/json
      description: Add menu by id Restaurant
      parameters:
      - description: the body to create a menu restaurant
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.MenuRestaurantInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: HTTP-X-UID. Fill with id user
        in: header
        name: HTTP-X-UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Create Menus
      tags:
      - Menus
  /create/orders:
    post:
      consumes:
      - application/json
      description: Create order user by resto
      parameters:
      - description: Account ID
        in: path
        name: Body
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/models.OrderHistory'
            type: object
      summary: Create order
      tags:
      - User
  /delete/menus:
    delete:
      consumes:
      - application/json
      description: Delete menus by restaurant id
      parameters:
      - description: menus id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: HTTP-X-UID. Fill with id user
        in: header
        name: HTTP-X-UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete Menus
      tags:
      - menus
  /get_all_resto:
    get:
      consumes:
      - application/json
      description: Get data for all resto
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get All Restaurant
      tags:
      - restaurant
  /login:
    post:
      consumes:
      - application/json
      description: Logging to get jwt token
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login User
      tags:
      - Auth
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - Auth
  /restaurant/create:
    post:
      consumes:
      - application/json
      description: Create Restaurant
      parameters:
      - description: the body to create a restaurant
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RestaurantInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: HTTP-X-UID. Fill with id user
        in: header
        name: HTTP-X-UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/models.Restaurant'
            type: object
      security:
      - BearerToken: []
      summary: Create data Restaurant
      tags:
      - restaurant
  /show/order/:id/restaurant:
    get:
      consumes:
      - application/json
      description: Get Order by id
      parameters:
      - description: Restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderHistory'
      summary: Show Order By Resto
      tags:
      - Order User
swagger: "2.0"